Added optional feature to prevent windows from overlapping
When a notification is spawned near one of the 4 corners corner it will
be moved to not overlap with the last notification intended to be
spawned near that specific corner. If the corner was the TOP_LEFT or
TOP_RIGHT then the notification will be below the previous or above it
otherwise. If the rearrangement will cause the notification to be off
screen vertically, it instead will be shifted horizontally from its
intended position into a new column.